Implement `RecognizedShape.interpretSvg`
